+++ SConstruct
@@ -3,4 +3,4 @@ SConscript('board/SConscript')
 
 # test files
 SConscript('tests/safety/SConscript')
-#SConscript('tests/usbprotocol/SConscript')
+SConscript('tests/usbprotocol/SConscript')
