+++ board/main.c
@@ -25,6 +25,7 @@
 #include "obj/gitversion.h"
 
 #include "main_comms.h"
+#include "can_comms.h"
 
 
 // ********************* Serial debugging *********************
